drop database after tests
